# api/docs/swagger.yaml
definitions:
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  object_builder_service.ActionRelation:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  object_builder_service.AutoFilter:
    properties:
      field_from:
        type: string
      field_to:
        type: string
    type: object
...
  object_builder_service.Cascading:
    properties:
      field_slug:
        type: string
      label:
        type: string
      order:
        type: integer
      table_slug:
        type: string
    type: object
...
  object_builder_service.CreateRelationRequest:
    properties:
      action_relations:
        items:
          $ref: '#/definitions/object_builder_service.ActionRelation'
        type: array
      auto_filters:
        items:
          $ref: '#/definitions/object_builder_service.AutoFilter'
        type: array
      cascading_tree_field_slug:
        type: string
      cascading_tree_table_slug:
        type: string
      cascadings:
        items:
          $ref: '#/definitions/object_builder_service.Cascading'
        type: array
      columns:
        items:
          type: string
        type: array
      default_limit:
        type: string
      default_values:
        items:
          type: string
        type: array
      dynamic_tables:
        items:
          $ref: '#/definitions/object_builder_service.DynamicTable'
        type: array
      editable:
        type: boolean
      group_fields:
        items:
          type: string
        type: array
      is_editable:
        type: boolean
      is_user_id_default:
        type: boolean
      multiple_insert:
        type: boolean
      multiple_insert_field:
        type: string
      object_id_from_jwt:
        type: boolean
      project_id:
        type: string
      quick_filters:
        items:
          $ref: '#/definitions/object_builder_service.QuickFilter'
        type: array
      relation_field_slug:
        type: string
      relation_table_slug:
        type: string
      summaries:
        items:
          $ref: '#/definitions/object_builder_service.Summary'
        type: array
      table_from:
        type: string
      table_to:
        type: string
      title:
        type: string
      type:
        type: string
      updated_fields:
        items:
          type: string
        type: array
      view_fields:
        items:
          type: string
        type: array
      view_type:
        type: string
    type: object
...
  object_builder_service.DynamicTable:
    properties:
      table_slug:
        type: string
      view_fields:
        items:
          type: string
        type: array
    type: object
...
  object_builder_service.QuickFilter:
    properties:
      default_value:
        type: string
      field_id:
        type: string
    type: object
...
  object_builder_service.Summary:
    properties:
      field_name:
        type: string
      formula_name:
        type: string
    type: object
...
  object_builder_service.UpdateRelationRequest:
    properties:
      action_relations:
        items:
          $ref: '#/definitions/object_builder_service.ActionRelation'
        type: array
      auto_filters:
        items:
          $ref: '#/definitions/object_builder_service.AutoFilter'
        type: array
      cascading_tree_field_slug:
        type: string
      cascading_tree_table_slug:
        type: string
      cascadings:
        items:
          $ref: '#/definitions/object_builder_service.Cascading'
        type: array
      columns:
        items:
          type: string
        type: array
      default_limit:
        type: string
      default_values:
        items:
          type: string
        type: array
      dynamic_tables:
        items:
          $ref: '#/definitions/object_builder_service.DynamicTable'
        type: array
      editable:
        type: boolean
      group_fields:
        items:
          type: string
        type: array
      id:
        type: string
      is_editable:
        type: boolean
      is_user_id_default:
        type: boolean
      multiple_insert:
        type: boolean
      multiple_insert_field:
        type: string
      object_id_from_jwt:
        type: boolean
      project_id:
        type: string
      quick_filters:
        items:
          $ref: '#/definitions/object_builder_service.QuickFilter'
        type: array
      relation_field_slug:
        type: string
      relation_table_slug:
        type: string
      summaries:
        items:
          $ref: '#/definitions/object_builder_service.Summary'
        type: array
      table_from:
        type: string
      table_to:
        type: string
      title:
        type: string
      type:
        type: string
      updated_fields:
        items:
          type: string
        type: array
      view_fields:
        items:
          type: string
        type: array
      view_type:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://udevs.io
paths:
...
  /v1/get-relation-cascading/{table_slug}:
    get:
      consumes:
      - application/json
      description: Get all relations
      operationId: get_relation_cascaders
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: table_slug
        in: path
        name: table_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: CascaderBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Get all relations
      tags:
      - Relation
...
  /v1/relation:
    get:
      consumes:
      - application/json
      description: Get all relations
      operationId: get_all_relations
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: project_id
        type: string
      - in: query
        name: relation_table_slug
        type: string
      - in: query
        name: table_id
        type: string
      - in: query
        name: table_slug
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: RelationBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Get all relations
      tags:
      - Relation
    post:
      consumes:
      - application/json
      description: Create relation
      operationId: create_relation
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: CreateRelationRequestBody
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/object_builder_service.CreateRelationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Relation data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create relation
      tags:
      - Relation
    put:
      consumes:
      - application/json
      description: Update relation
      operationId: update_relation
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: UpdateRelationRequestBody
        in: body
        name: relation
        required: true
        schema:
          $ref: '#/definitions/object_builder_service.UpdateRelationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Relation data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Update relation
      tags:
      - Relation
  /v1/relation/{relation_id}:
    delete:
      consumes:
      - application/json
      description: Delete Relation
      operationId: delete_relation
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: relation_id
        in: path
        name: relation_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Delete Relation
      tags:
      - Relation
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"